passphrase_jenny/requirements.txt
pyperclip==1.9.0
python-dotenv==1.0.1
